fix(log): resolve-audit stderr echo is opt-in; adi_inspect -Werror
Two repairs to keep the clang -Werror build and the parity gates green
after the v1.8.73-88 batch:

1. The new RESOLVED-line audit wrote to raw stderr whenever no explicit
   console sink was set - i.e. in every embedded consumer of the DLL
   (php_openads under Apache, Harbour apps, the qa-diff harness). SAP's
   ace64.dll never writes to the host process's streams, and the noise
   dropped the parity gates from 41/1+37/18 to 9/33+5/50 (every OA
   output gained audit lines SAP doesn't emit). The stderr fallback now
   requires OPENADS_LOG or OPENADS_RESOLVE_VERBOSE; an explicit
   set_audit_console() sink and the OPENADS_LOG_FILE sink behave exactly
   as before, so serverd/debug workflows keep their lines by setting the
   env or the sink.

2. tools/adi_inspect failed clang -Werror (implicit int->u16 in u16_le,
   long/unsigned mixes in the two fseek offsets) - MSVC doesn't warn on
   these. Same class of fix as 16eb3657.

Both gates back to pmsys 41/1, mp 37/18.

+++ tools/adi_inspect/main.cpp
@@ -25,7 +25,7 @@ constexpr std::uint16_t LVL_TAGDIR = 3;
 using Page = std::array<std::uint8_t, PAGE_SIZE>;
 
 std::uint16_t u16_le(const std::uint8_t* p) {
-    return static_cast<std::uint16_t>(p[0]) | (static_cast<std::uint16_t>(p[1]) << 8);
+    return static_cast<std::uint16_t>(p[0] | (p[1] << 8));
 }
 
 std::uint32_t u32_le(const std::uint8_t* p) {
@@ -52,7 +52,7 @@ struct TagInfo {
 
 std::string read_field_name_from_page(FILE* f, std::uint32_t page_no) {
     Page pg;
-    std::fseek(f, static_cast<long>(page_no) * PAGE_SIZE, SEEK_SET);
+    std::fseek(f, static_cast<long>(page_no * PAGE_SIZE), SEEK_SET);
     if (std::fread(pg.data(), 1, PAGE_SIZE, f) != PAGE_SIZE) return "";
     
     // Field marker page: field numbers encoded as 1-byte each, terminated by 0
@@ -132,7 +132,7 @@ void inspect_adi(const fs::path& path) {
 
         // Read per-tag header page for unique flag and v2 metadata
         Page hdr_pg;
-        std::fseek(f, static_cast<long>(xx) * PAGE_SIZE, SEEK_SET);
+        std::fseek(f, static_cast<long>(xx * PAGE_SIZE), SEEK_SET);
         if (std::fread(hdr_pg.data(), 1, PAGE_SIZE, f) == PAGE_SIZE) {
             tag.unique = (hdr_pg[14] & 0x01u) != 0;
